refactor: rename file to match provided functionality
